+++ src/cfn-binder/cfn-task-runner.ts
@@ -5,11 +5,14 @@ import { ICfnTask } from './cfn-task-provider';
 
 export class CfnTaskRunner {
 
-    public static async RunTasks(tasks: ICfnTask[], stackName: string) {
+    public static async RunTasks(tasks: ICfnTask[], stackName: string, maxConcurrentTasks: number, failedTasksTolerance: number) {
 
         const delegate: ITaskRunnerDelegates<ICfnTask> = {
             onTaskRanFailed: (task, err) => {
                 ConsoleUtil.LogError(`failed executing stack ${task.stackName} in account ${task.accountId} (${task.region}). Reason: ${err}`);
+            },
+            onTaskSkippedBecauseDependencyFailed: (task) => {
+                ConsoleUtil.LogError(`skip executing stack ${task.stackName} in account ${task.accountId} (${task.region}). Reason: dependency has failed.`);
             },
             onTaskRanSuccessfully: (task) => {
                 ConsoleUtil.LogInfo(`stack ${task.stackName} successfully ${task.action === 'Delete' ? 'deleted from' : 'updated in' } ${task.accountId}/${task.region}.`);
@@ -19,8 +22,11 @@ export class CfnTaskRunner {
                 throw new OrgFormationError(`circular dependency on stack ${stackName} for targets ${targets.join(', ')}`);
              },
             throwDependencyOnSelfException: (task) => {throw new OrgFormationError(`stack ${task.stackName} has dependency on self target account ${task.accountId} / ${task.region}`); },
-            maxConcurrentTasks: 10,
-            failedTasksTolerance: 1,
+            onFailureToleranceExceeded: (totalTasksFailed: number, tolerance: number) => {
+                throw new OrgFormationError(`number failed stacks ${totalTasksFailed} exceeded tolerance for failed stacks ${tolerance}`);
+            },
+            maxConcurrentTasks,
+            failedTasksTolerance,
         };
         await GenericTaskRunner.RunTasks<ICfnTask>(tasks, delegate);
     }
@@ -32,6 +38,10 @@ export class CfnTaskRunner {
                 const sname = stackName ? `stack ${stackName} ` : '';
                 ConsoleUtil.LogError(`unable to validate template for ${sname}account ${task.accountId} (${task.region}). Reason: ${err}`);
             },
+            onTaskSkippedBecauseDependencyFailed: (task) => {
+                const sname = stackName ? `stack ${stackName} ` : '';
+                ConsoleUtil.LogError(`unable to validate template for ${sname}account ${task.accountId} (${task.region}). Reason: dependency has failed.`);
+            },
             onTaskRanSuccessfully: (task) => {
                 const sname = stackName ? `stack ${stackName} ` : '';
                 ConsoleUtil.LogInfo(`template for ${sname}account ${task.accountId}/${task.region} valid.`);
@@ -39,6 +49,9 @@ export class CfnTaskRunner {
             throwCircularDependency: (ts) => {
                 const targets = ts.map((x) => x.accountId + '/' + x.region);
                 throw new OrgFormationError(`circular dependency for targets ${targets.join(', ')}`);
+             },
+             onFailureToleranceExceeded: (totalTasksFailed: number, tolerance: number) => {
+                throw new OrgFormationError(`number failed tasks ${totalTasksFailed} exceeded tolerance for failed tasks ${tolerance}`);
              },
             throwDependencyOnSelfException: (task) => {throw new OrgFormationError(`template has dependency on self target account ${task.accountId} / ${task.region}`); },
             maxConcurrentTasks: 99,