+++ src/file-util.ts
@@ -1,7 +1,0 @@
-import { writeFileSync } from 'fs';
-
-export class FileUtil {
-    public static writeFileSync(path: string, contents: string) {
-        writeFileSync(path, contents);
-    }
-}